MOTECH-2726 Services moved to tasks package
