+++ commcare/src/main/java/org/motechproject/commcare/tasks/builder/CaseActionBuilder.java
@@ -27,7 +27,7 @@ import static org.motechproject.tasks.domain.mds.ParameterType.UNICODE;
  */
 public class CaseActionBuilder implements ActionBuilder {
 
-    private static final String ACTION_CASE_SERVICE = "org.motechproject.commcare.service.CaseActionService";
+    private static final String ACTION_CASE_SERVICE = "org.motechproject.commcare.tasks.CaseActionService";
 
     private CommcareConfigService configService;
