+++ commcare/src/test/java/org/motechproject/commcare/events/imports/ImportFormActionEventHandlerTest.java
@@ -3,14 +3,13 @@ package org.motechproject.commcare.events.imports;
 import org.joda.time.DateTime;
 import org.junit.Before;
 import org.junit.Test;
-import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.motechproject.commcare.events.constants.EventDataKeys;
 import org.motechproject.commcare.events.constants.EventSubjects;
 import org.motechproject.commcare.pull.CommcareFormImporterImpl;
 import org.motechproject.commcare.pull.CommcareTasksFormImporterFactory;
 import org.motechproject.commcare.service.CommcareFormService;
-import org.motechproject.commcare.service.impl.imports.ImportFormActionServiceImpl;
+import org.motechproject.commcare.tasks.imports.ImportFormActionServiceImpl;
 import org.motechproject.commons.api.Range;
 import org.motechproject.event.MotechEvent;
 import org.motechproject.event.listener.EventRelay;
